feat: rendering dots on client

+++ web/2/src/main/scala/com/itmo/mrdvd/dto/Dot.scala
@@ -1,3 +1,3 @@
 package com.itmo.mrdvd.dto
 
-class Dot(val X: Int, val Y: Float, val R: Float)
+class Dot(val X: Float, val Y: Float, val R: Float)
